StudyByte/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
nltk==3.8.1
textblob==0.17.1
PyPDF2==3.0.1
transformers==4.33.0
torch==2.0.1
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0

# AI Model APIs
openai==1.12.0
google-generativeai==0.5.4

# Video Processing for Smart Summarizer
openai-whisper
moviepy

# Video Processing & Speech-to-Text
openai-whisper==20240930
moviepy==1.0.3
ffmpeg-python==0.2.0